display group name in filter pills everywhere
