+++ frontend/src/components/filter/activityReportFilters.js
@@ -26,6 +26,7 @@ import FilterTTAType, { displayTtaTypeQuery } from './FilterTTAType';
 import MyReportsSelect from './MyReportsSelect';
 import FilterGroups from './FilterGroups';
 import FilterDeliveryMethod from './FilterDeliveryMethod';
+import { useDisplayGroups } from './utils';
 
 const EMPTY_MULTI_SELECT = {
   is: [],
@@ -439,7 +440,7 @@ export const groupsFilter = {
   display: 'Group',
   conditions: FILTER_CONDITIONS,
   defaultValues: EMPTY_MULTI_SELECT,
-  displayQuery: handleArrayQuery,
+  displayQuery: useDisplayGroups,
   renderInput: (id, condition, query, onApplyQuery) => (
     <FilterGroups
       inputId={`group-${condition}-${id}`}
