+++ frontend/src/components/filter/grantFilters.js
@@ -1,21 +1,8 @@
 /* eslint-disable import/prefer-default-export */
-import React, { useContext } from 'react';
+import React from 'react';
 import { FILTER_CONDITIONS, EMPTY_MULTI_SELECT } from '../../Constants';
 import FilterGroups from './FilterGroups';
-import { MyGroupsContext } from '../MyGroupsProvider';
-
-const useDisplayGroups = (query) => {
-  const { myGroups } = useContext(MyGroupsContext);
-
-  if (!query || query.length === 0) {
-    return '';
-  }
-
-  return [query].flat().map((q) => {
-    const group = myGroups.find((g) => g.id === q);
-    return group ? group.name : '';
-  }).join(', ');
-};
+import { useDisplayGroups } from './utils';
 
 export const groupsFilter = {
   id: 'group',
